+++ src/app/dataset-flow/dataset-flow-details/tabs/flow-details-history-tab/flow-details-history-tab.helpers.spec.ts
@@ -18,7 +18,11 @@ import {
     mockHistoryFragmentWithFinishedStatus,
 } from "./flow-details-history-tab.helpers.mock";
 import { FlowStatus } from "src/app/api/kamu.graphql.interface";
-import { mockDatasetExecuteTransformFlowSummaryData } from "src/app/dataset-flow/flows-table/flows-table.helpers.mock";
+import {
+    mockDatasetExecuteTransformFlowDescriptionUpdateResultUnknown,
+    mockDatasetExecuteTransformFlowSummaryData,
+    mockDatasetPollingIngestFlowDescriptionUpdateResultUnknown,
+} from "src/app/dataset-flow/flows-table/flows-table.helpers.mock";
 import timekeeper from "timekeeper";
 
 describe("DatasetFlowDetailsHelpers", () => {
@@ -115,6 +119,24 @@ describe("DatasetFlowDetailsHelpers", () => {
                 mockDatasetExecuteTransformFlowSummaryData,
             ),
         ).toEqual(flowEventSubMessageResults[14]);
+    });
+
+    it(`should check flow event submessage with typename = FlowDescriptionUpdateResultUnknown and flow outcome = Success (ExecuteTransform)`, () => {
+        expect(
+            DatasetFlowDetailsHelpers.flowEventSubMessage(
+                mockHistoryFragmentWithFinishedStatus,
+                mockDatasetExecuteTransformFlowDescriptionUpdateResultUnknown,
+            ),
+        ).toEqual(flowEventSubMessageResults[15]);
+    });
+
+    it(`should check flow event submessage with typename = FlowDescriptionUpdateResultUnknown and flow outcome = Success (FlowDescriptionDatasetPollingIngest)`, () => {
+        expect(
+            DatasetFlowDetailsHelpers.flowEventSubMessage(
+                mockHistoryFragmentWithFinishedStatus,
+                mockDatasetPollingIngestFlowDescriptionUpdateResultUnknown,
+            ),
+        ).toEqual(flowEventSubMessageResults[15]);
     });
 
     it(`should check don't show dynamic image`, () => {