+++ CHANGELOG.md
@@ -4,6 +4,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+### Added
+- Transform settings: enable a transform without setting batching parameters
+
 ## [0.43.0] - 2025-03-25
 ### Added
 - Flows table: added a hint to a long message