+++ CHANGELOG.md
@@ -4,6 +4,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+### Fixed
+- Wrong redirect after login
+### Added
+- Implemented `MergeStrategySnapshot` and `MergeStrategyUpsertStream` merge strategies
+- Implemented unknown response for broken dataset flows
+
+
 ## [0.43.0] - 2025-03-25
 ### Added
 - Flows table: added a hint to a long message