+++ CHANGELOG.md
@@ -7,6 +7,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 ### Added
 - Transform settings: enable a transform without setting batching parameters
+- Implemented `MergeStrategySnapshot` and `MergeStrategyUpsertStream` merge strategies
+- Implemented unknown response for broken dataset flows
+### Fixed
+- Wrong redirect after login
+
 
 ## [0.43.0] - 2025-03-25
 ### Added