+++ CHANGELOG.md
@@ -4,7 +4,7 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [Unreleased]
+## [0.44.0] - 2025-03-31
 ### Added
 - Transform settings: enable a transform without setting batching parameters
 - Implemented `MergeStrategySnapshot` and `MergeStrategyUpsertStream` merge strategies